set_mac needs to work for full 48-bit MAC addresses so that the packets sent by the server actually reach the client

+++ l2fwd/util.h
@@ -66,7 +66,7 @@ inline uint32_t fastrand(uint64_t* seed);
 int bitcount(int n);
 int *get_active_bits(int mask);
 
-inline void set_mac(uint8_t *mac_ptr, uint32_t mac_addr);
+inline void set_mac(uint8_t *mac_ptr, unsigned long long mac_addr);
 inline uint32_t get_mac(uint8_t *mac_ptr);
 inline void swap_mac(uint8_t *src_mac_ptr, uint8_t *dst_mac_ptr);
